requirements.txt
Flask==2.0.1
Flask-Session==0.4.0
Werkzeug==2.0.1
Flask-SQLAlchemy==2.5
SQLAlchemy==1.4.23
requests==2.26.0
python-dotenv==0.19.1
